+++ templates/single-pw_event.php
@@ -21,10 +21,6 @@ while ( have_posts() ) :
 	pw_get_template_part( 'single-event/key-details-strip', '', [ 'post_id' => $pw_event_post_id ] );
 	do_action( 'pw_after_event_key_details_strip', $pw_event_post_id );
 
-	do_action( 'pw_before_event_programme', $pw_event_post_id );
-	pw_get_template_part( 'single-event/programme', '', [ 'post_id' => $pw_event_post_id ] );
-	do_action( 'pw_after_event_programme', $pw_event_post_id );
-
 	do_action( 'pw_before_event_ticket_cta', $pw_event_post_id );
 	pw_get_template_part( 'single-event/ticket-cta', '', [ 'post_id' => $pw_event_post_id ] );
 	do_action( 'pw_after_event_ticket_cta', $pw_event_post_id );
@@ -32,10 +28,6 @@ while ( have_posts() ) :
 	do_action( 'pw_before_event_venue_details', $pw_event_post_id );
 	pw_get_template_part( 'single-event/venue-details', '', [ 'post_id' => $pw_event_post_id ] );
 	do_action( 'pw_after_event_venue_details', $pw_event_post_id );
-
-	do_action( 'pw_before_event_add_ons', $pw_event_post_id );
-	pw_get_template_part( 'single-event/add-ons', '', [ 'post_id' => $pw_event_post_id ] );
-	do_action( 'pw_after_event_add_ons', $pw_event_post_id );
 
 	do_action( 'pw_before_event_cta', $pw_event_post_id );
 	pw_get_template_part( 'single-event/cta', '', [ 'post_id' => $pw_event_post_id ] );