+++ templates/single-pw_meeting_room.php
@@ -25,10 +25,6 @@ while ( have_posts() ) :
 	pw_get_template_part( 'single-meeting-room/description', '', [ 'post_id' => $pw_meeting_room_post_id ] );
 	do_action( 'pw_after_meeting_room_description', $pw_meeting_room_post_id );
 
-	do_action( 'pw_before_meeting_room_catering_note', $pw_meeting_room_post_id );
-	pw_get_template_part( 'single-meeting-room/catering-note', '', [ 'post_id' => $pw_meeting_room_post_id ] );
-	do_action( 'pw_after_meeting_room_catering_note', $pw_meeting_room_post_id );
-
 	do_action( 'pw_before_meeting_room_floor_plan', $pw_meeting_room_post_id );
 	pw_get_template_part( 'single-meeting-room/floor-plan', '', [ 'post_id' => $pw_meeting_room_post_id ] );
 	do_action( 'pw_after_meeting_room_floor_plan', $pw_meeting_room_post_id );